Progress on series race.

+++ src/test/java/common/RaceTest.java
@@ -3,74 +3,49 @@ package common;
 import org.junit.jupiter.api.AfterEach;
 import uk.ac.standrews.cs.utilities.FileManipulation;
 
-import java.io.FileInputStream;
 import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
+import java.nio.file.*;
+import java.nio.file.attribute.BasicFileAttributes;
 import java.util.*;
 import java.util.stream.Stream;
 
+import static java.nio.file.FileVisitResult.CONTINUE;
 import static org.junit.jupiter.api.Assertions.*;
 
 public abstract class RaceTest {
 
-    private Properties properties;
-    private Path resources_expected_outputs;
+    private Path config_file_path;
+    private Path expected_output_directory;
     private Path temp_directory;
-    private Path temp_output_sub_directory;
+    private Path temp_output_directory;
 
-    protected abstract Race makeRace(Properties properties) throws IOException;
+    // When enabled, test results are retained on the desktop.
+    private static final boolean DEBUG = false;
+    private static final String DEBUG_FILES_LOCATION = "/Users/gnck/Desktop/temp";
+
+    protected abstract Race makeRace(Path config_file_path) throws IOException;
     protected abstract String getResourcesPath();
 
     @AfterEach
     public void tearDown() throws IOException {
 
-        // Disable this when debugging and you don't want the test results to be immediately deleted.
-
-        FileManipulation.deleteDirectory(temp_directory);
+        if (!DEBUG) FileManipulation.deleteDirectory(temp_directory);
     }
 
     protected void configureTest(final String test_resource_root) throws IOException {
 
-        // Swap these when debugging and you don't want the test results to be immediately deleted.
+        temp_directory = DEBUG ? Paths.get(DEBUG_FILES_LOCATION) : Files.createTempDirectory("temp");
 
-        //temp_directory = Files.createTempDirectory(null);
-        temp_directory = Paths.get("/Users/gnck/Desktop/temp");
-
-        final Path temp_input_sub_directory = Files.createDirectories(temp_directory.resolve("input"));
-        temp_output_sub_directory = Files.createDirectories(temp_directory.resolve("output"));
-
-        final Path resources_root = Paths.get("src/test/resources/" + getResourcesPath() + test_resource_root);
+        final Path resources_root = Paths.get("src/test/resources/").resolve(getResourcesPath()).resolve(test_resource_root);
         final Path resources_inputs = resources_root.resolve("input");
-        resources_expected_outputs = resources_root.resolve("expected");
-        final Path resources_config = resources_inputs.resolve("config.txt");
+        expected_output_directory = resources_root.resolve("expected");
 
-        copyFilesBetweenDirectories(resources_inputs, temp_input_sub_directory);
+        temp_output_directory = Files.createDirectories(temp_directory.resolve("output"));
 
-        properties = getProperties(resources_config);
-        properties.setProperty("WORKING_DIRECTORY", temp_directory.toString());
-    }
+        final Path temp_input_directory = temp_directory.resolve("input");
+        config_file_path = temp_input_directory.resolve("config.txt");
 
-    private static void copyFilesBetweenDirectories(final Path source, final Path destination) throws IOException {
-
-        try (final Stream<Path> list = Files.list(source)) {
-
-            for (final Iterator<Path> iterator = list.iterator(); iterator.hasNext(); ) {
-                Path file = iterator.next();
-                Files.copy(file, destination.resolve(file.getFileName()));
-            }
-        }
-    }
-
-    private static Properties getProperties(final Path path) throws IOException {
-
-        try (final FileInputStream in = new FileInputStream(path.toFile())) {
-
-            Properties properties = new Properties();
-            properties.load(in);
-            return properties;
-        }
+        copyDirectory(resources_inputs, temp_input_directory);
     }
 
     public void testExpectedException(final String configuration_name, final String expected_error_message) throws Exception {
@@ -79,7 +54,7 @@ public abstract class RaceTest {
 
         RuntimeException thrown = assertThrows(
                 RuntimeException.class,
-                () -> makeRace(properties).processResults()
+                () -> makeRace(config_file_path).processResults()
         );
 
         assertEquals(expected_error_message, thrown.getMessage());
@@ -88,8 +63,8 @@ public abstract class RaceTest {
     protected void testExpectedCompletion(final String configuration_name) throws Exception {
 
         configureTest(configuration_name);
-        makeRace(properties).processResults();
-        assertThatDirectoryContainsAllExpectedContent(resources_expected_outputs, temp_output_sub_directory);
+        makeRace(config_file_path).processResults();
+        assertThatDirectoryContainsAllExpectedContent(expected_output_directory, temp_output_directory);
     }
 
     private static void assertThatDirectoryContainsAllExpectedContent(final Path expected, final Path actual) throws IOException {
@@ -146,7 +121,38 @@ public abstract class RaceTest {
     }
 
     private static boolean fileInExpectedDirectoryShouldBeIgnored(final String file_name) {
-
         return file_name.equals(".DS_Store");
+    }
+
+    private static void copyDirectory(final Path source_directory, final Path destination_directory) throws IOException {
+
+        final FileVisitor<Path> copier = new FileVisitor<>() {
+
+            @Override
+            public FileVisitResult preVisitDirectory(Path dir, BasicFileAttributes attrs) throws IOException {
+
+                Files.copy(dir, destination_directory.resolve(source_directory.relativize(dir)));
+                return CONTINUE;
+            }
+
+            @Override
+            public FileVisitResult visitFile(Path file, BasicFileAttributes attrs) throws IOException {
+
+                Files.copy(file, destination_directory.resolve(source_directory.relativize(file)));
+                return CONTINUE;
+            }
+
+            @Override
+            public FileVisitResult visitFileFailed(Path file, IOException exc) {
+                return CONTINUE;
+            }
+
+            @Override
+            public FileVisitResult postVisitDirectory(Path dir, IOException exc) {
+                return CONTINUE;
+            }
+        };
+
+        Files.walkFileTree(source_directory, copier);
     }
 }
